code updates

+++ twin-stick-shooter-game-1/constants.h
@@ -1,6 +1,6 @@
 #pragma once
 #define _USE_MATH_DEFINES
-#include <math.h>
+#include <math.h> //can't be <cmath> for some reason
 
 const int TICK_CALLBACK = 10;
 //16ms = 60Hz, 20ms = 50Hz, 10ms = 100Hz
@@ -8,5 +8,5 @@ const int TICK_CALLBACK = 10;
 const float GAME_WIDTH = 1000;
 const float GAME_HEIGHT = 1000;
 
-const double PI = M_PI;
+const float PI = static_cast<float>(M_PI);
